Fixed use variables with function variables #1427
