+++ docs/CHANGELOG-v1.md
@@ -24,6 +24,9 @@ What's changed since v1.15.2:
 - Engineering:
   - Bump Microsoft.NET.Test.Sdk to v17.2.0.
     [#1410](https://github.com/Azure/PSRule.Rules.Azure/pull/1410)
+- Bug fixes:
+  - Fixed `Azure.Template.UseVariables` does not accept function variables names by @BernieWhite.
+    [#1427](https://github.com/Azure/PSRule.Rules.Azure/issues/1427)
 
 ## v1.15.2
