+++ tests/PSRule.Rules.Azure.Tests/Tests.Bicep.6.content.json
@@ -1,0 +1,3 @@
+{
+    "name": "value"
+}
